Deleted unnecessary file, added code to the main.py gui
